+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.3
+
+- Adds parameter to skip a failing migration and update the database version
+
+## 1.0.2
+
+- Adds function to reset database version
+
 ## 1.0.1+2
 
 - Bumps get_it version to 5.x.x
@@ -8,7 +16,7 @@
 
 ## 1.0.1
 
-- Updates _DatabaseVersionKey value. 
+- Updates \_DatabaseVersionKey value.
 
 ## 1.0.0 - First release